Questionクラスを削除
