+++ main.js
@@ -2,16 +2,16 @@
 
 import { exec } from "node:child_process";
 import { Client } from "./client.js";
-import { Question } from "./question.js";
+import { enterApiToken, askSearchParams, selectSchool, confirmMapDisplay } from "./questions.js";
 
 let token;
 if (process.env.GAKKOU_SEARCH_API_TOKEN) {
   token = process.env.GAKKOU_SEARCH_API_TOKEN;
 } else {
-  token = await Question.enterApiToken();
+  token = await enterApiToken();
 }
 
-const searchParams = await Question.askSearchParams();
+const searchParams = await askSearchParams();
 const params = {};
 Object.keys(searchParams).forEach((key) => {
   if (searchParams[key] !== "0") {
@@ -29,9 +29,7 @@ try {
   } else {
     console.error("学校データの取得に失敗しました。");
   }
-  console.error(
-    `Error: ${error.message}, Status: ${error.response.status}, Status Text: ${error.response.statusText}`
-  );
+  console.error(`Error: ${error.message}, Status: ${error.response.status}, Status Text: ${error.response.statusText}`);
   process.exit();
 }
 
@@ -40,19 +38,17 @@ if (total === 0) {
   console.log("学校が見つかりませんでした");
   process.exit();
 } else if (total > 100) {
-  console.log(
-    `\n${total}件の学校が見つかりました。最初の100件のみ表示します。\n`
-  );
+  console.log(`\n${total}件の学校が見つかりました。最初の100件のみ表示します。\n`);
 } else {
   console.log(`\n${total}件の学校が見つかりました。\n`);
 }
 
 const schools = response.schools;
-const selected = await Question.selectSchool(schools);
+const selected = await selectSchool(schools);
 const selectedSchool = selected.school;
 console.log(selectedSchool.info());
 
-const mapConfirm = await Question.confirmMapDisplay();
+const mapConfirm = await confirmMapDisplay();
 if (mapConfirm) {
   const mapUrl = selectedSchool.mapUrl();
   exec(`open "${mapUrl}"`, (error) => {
